pyproject.toml
[project]
dependencies = [
  "django-q2",
  "djangorestframework",
  "djangorestframework-simplejwt",
  "django-cors-headers",
  "django~=5.2",
  "gunicorn",
  "psycopg2",
]
description = "A Pick 20 Pool website for NCAA March Madness"
name = "pick20-pool"
readme = "README.md"
requires-python = "==3.11.*"
version = "0.1.0"

[dependency-groups]
dev = ["bump-my-version>=1.1.2"]

[[tool.uv.index]]
default = true
name = "pypi"
url = "https://pypi.org/simple"

[tool.ruff]
exclude = ["**/*__init__.py", "**/migrations/*", ".venv/*", "build/*", "dist/*"]
fix = true
line-length = 120
src = ["pick20/"]

[tool.ruff.lint]
extend-select = ["I"]
ignore = ["C901", "E203", "E731", "F403", "F405"]


[tool.bumpversion]
commit = true
current_version = "0.1.0"
parse = """(?x)
    (?P<major>0|[1-9]\\d*)\\.
    (?P<minor>0|[1-9]\\d*)\\.
    (?P<patch>0|[1-9]\\d*)
"""
# when bumping the version, uv.lock immediately becomes out of date. update local package only
pre_commit_hooks = ["uv lock --upgrade-package pick20-pool; git add uv.lock"]
tag = true

serialize = ["{major}.{minor}.{patch}"]

[[tool.bumpversion.files]]
filename = "pyproject.toml"
replace = 'version = "{new_version}"'
search = 'version = "{current_version}"'

[[tool.bumpversion.files]]
filename = "pick20/pick20/__init__.py"
replace = '__version__ = "{new_version}"'
search = '__version__ = "{current_version}"'

[tool.pytest.ini_options]
django_find_project = false
pythonpath = '. pick20'
